+++ components/script/Cargo.toml
@@ -33,6 +33,7 @@ serde_json = "1.0"
 accountable-refcell = { version = "0.2.0", optional = true }
 app_units = "0.7"
 arrayvec = "0.7"
+atomic_refcell = "0.1"
 backtrace = "0.3"
 base64 = "0.10.1"
 bitflags = "1.0"
@@ -54,6 +55,7 @@ enum-iterator = "0.3"
 euclid = "0.22"
 fnv = "1.0"
 fxhash = "0.2"
+gfx_traits = { path = "../gfx_traits" }
 headers = "0.3"
 html5ever = "0.26"
 http = "0.2"
@@ -83,6 +85,7 @@ percent-encoding = "2.0"
 phf = "0.8"
 pixels = { path = "../pixels" }
 profile_traits = { path = "../profile_traits" }
+range = { path = "../range" }
 ref_filter_map = "1.0.1"
 regex = "1.1"
 script_layout_interface = { path = "../script_layout_interface" }
